% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COCAcaller.R
\name{COCAcaller}
\alias{COCAcaller}
\title{Identify COCA Subtype from miRNA Expression in Chronic Pancreatitis}
\usage{
COCAcaller(data)
}
\arguments{
\item{data}{A miRNA expression matrix or dataframe with sample rows and gene columns. Each row represents a sample and each column represents a gene.}
}
\value{
A dataframe with the predicted probabilities for each COCA subtype and the normal category. Additionally, it includes a column indicating the predicted group (Normal, COCA1, COCA2, COCA3) with the highest probability for each sample.
}
\description{
This function identifies the COCA subtype from miRNA expression data in chronic pancreatitis patients. It uses a pre-trained multi-layer perceptron classifier (mlpcla) to predict the probability of each sample belonging to one of the COCA subtypes (COCA1, COCA2, COCA3) or being normal.
}
\details{
This function takes a miRNA expression dataset as input and applies a pre-trained multi-layer perceptron classifier (mlpcla) to predict the subtype of chronic pancreatitis for each sample. The function outputs a dataframe with the probabilities of each sample being classified as Normal, COCA1, COCA2, or COCA3, along with the predicted group with the highest probability. The classifier should be trained prior to using this function.
}
\author{
Zaoqu Liu; Email: liuzaoqu@163.com
}
